+++ R/multiqc.R
@@ -95,7 +95,7 @@ parse_metadata <- function(parsed, samples, find_metadata) {
 #' @param find_metadata A single function that will be called with a sample name and the
 #' parsed JSON for the entire report and returns a named list of metadata fields for the sample.
 #' Refer to the vignette for an example.
-#' @param sections A string vector of zero or more sections to include in the output. 
+#' @param sections A string vector of zero or more sections to include in the output.
 #' Each section can be:
 #' \describe{
 #' \item{"plot"}{Parse plot data. Note that you should also provide a list of plots via the `plots` argument}
@@ -103,28 +103,31 @@ parse_metadata <- function(parsed, samples, find_metadata) {
 #' \item{"raw"}{Parse the raw data section}
 #' }
 #' This defaults to 'general', which tends to contain the most useful statistics
-#' @param plots A string vector, each of which contains the ID of a plot you 
+#' @param plots A string vector, each of which contains the ID of a plot you
 #' want to include in the output. You can use [TidyMultiqc::list_plots()] to help here.
-#' @param plot_parsers. [Advanced] A named list of custom parser functions. 
+#' @param plot_parsers **Advanced**. A named list of custom parser functions.
 #' The names of the list should correspond to plotly plot types, such as "xy_line", and the values should be functions
 #' that return a named list of named lists. For the return value, the outer list is named by the sample ID, and the inner list
 #' is named by the name of the column. Refer to the source code for some examples.
 #' @export
 #' @return A tibble (data.frame subclass) with QC data and metadata as columns, and samples as rows.
-#' Columns are named according to the respective section they belong to, 
+#' Columns are named according to the respective section they belong to,
 #' and will always be listed in the following order:
-#' \item{`metadata.X`}{This column contains metadata for this sample. 
+#' \item{`metadata.X`}{This column contains metadata for this sample.
 #' By default this is only the sample ID, but if you have provided the
-#'`find_metadata` argument, there may be more columns.}
+#' `find_metadata` argument, there may be more columns.}
 #' \item{`general.X`}{This column contains a generally useful summary statistic for each sample}
-#' \item{`plot.X`}{This column contains a data frame of plot data for each sample}
-#' \item{`raw.X`}{This column contains a raw summary statistic or value relating to each sample}
-#' }
+#' \item{`plot.X`}{This column contains a data frame of plot data for each sample.
+#' Refer to the plot parsers documentation (ie the `parse_X` functions) for more information on the output format. }
+#' \item{`raw.X`}{This column contains a raw summary statistic or value relating to each sample }
+#' @seealso [TidyMultiqc::parse_xyline_plot()] [TidyMultiqc::parse_bar_graph()]
 #' @examples
 #' load_multiqc(system.file("extdata", "wgs/multiqc_data.json", package = "TidyMultiqc"))
 load_multiqc <- function(paths,
                          plots = NULL,
-                         find_metadata = function(...) { list() },
+                         find_metadata = function(...) {
+                           list()
+                         },
                          plot_parsers = list(),
                          sections = "general") {
   assertthat::assert_that(all(sections %in% c(
@@ -141,7 +144,7 @@ load_multiqc <- function(paths,
         purrr::map(~ switch(.,
           general = parse_general(parsed),
           raw = parse_raw(parsed),
-          plot = parse_plots(parsed, plots = plots, plot_parsers=plot_parsers)
+          plot = parse_plots(parsed, plots = plots, plot_parsers = plot_parsers)
         )) %>%
         purrr::reduce(~ purrr::list_merge(.x, !!!.y), .init = list()) %>%
         purrr::imap(~ purrr::list_merge(.x, metadata.sample_id = .y))
@@ -152,14 +155,16 @@ load_multiqc <- function(paths,
         dplyr::bind_rows()
     }) %>%
     # Only arrange the columns if we have at least 1 column
-      `if`(
+    `if`(
       # Move the columns into the order: metadata, general, plot, raw
       ncol(.) > 0,
       (.) %>%
         dplyr::relocate(dplyr::starts_with("raw")) %>%
         dplyr::relocate(dplyr::starts_with("plot")) %>%
         dplyr::relocate(dplyr::starts_with("general")) %>%
-        dplyr::relocate(dplyr::starts_with("metadata")),
+        dplyr::relocate(dplyr::starts_with("metadata")) %>%
+        # Always put the sample ID at the start
+        dplyr::relocate(metadata.sample_id),
       .
     )
 }