+++ R/plots.R
@@ -1,67 +1,6 @@
-# Internal plot parsing functions
+# Plot parsing functions
 
-#' Takes the JSON dictionary for an xyline plot, and returns a named list of 
-#' data frames, one for each sample. 
-#' @keywords internal
-#' @import rlang
-#' @noRd
-parse_xyline_plot <- function(plot_data, name) {
-  # This only works on xyline plots
-  assertthat::assert_that(plot_data$plot_type == "xy_line")
-  
-  plot_data$datasets %>%
-    purrr::map(function(dataset) {
-      # For some reason there are two levels of nesting here
-      dataset %>%
-        kv_map(function(subdataset) {
-          name = stringr::str_c("plot", name, sep=".")
-          list(
-            key = subdataset$name,
-            value = subdataset$data %>%
-              purrr::map_dfr(~list(x=.[[1]], y=.[[2]])) %>%
-              # Chop the multi-row data frame into one row
-              tidyr::nest({{name}} := tidyr::everything()) #%>%
-          )
-        })
-    }) %>%
-    purrr::reduce(~ purrr::list_merge(.x, !!!.y))
-}
-
-#' Takes the JSON dictionary for a bar graph, and returns a named list of 
-#' data frames, one for each sample. 
-#' @keywords internal
-#' @import rlang
-#' @noRd
-parse_bar_graph <- function(plot_data, name) {
-  # This only works on bar_graphs
-  assertthat::assert_that(plot_data$plot_type == "bar_graph")
-
-  # Make a list of samples
-  samples <- plot_data$samples[[1]] %>% purrr::flatten_chr()
-  
-  colname = stringr::str_c("plot", sanitise_column_name(name), sep = ".") 
-  
-  plot_data$datasets[[1]] %>%
-    # First, build up a dictionary of samples -> dictionary of quality metrics
-    purrr::map(function(dataset) {
-      segment_name <- dataset$name
-      dataset$data %>%
-        # For this segment, each sample has a value
-        kv_map(function(value, idx) {
-          list(
-            key = samples[[idx]],
-            value = list(value) %>% purrr::set_names(sanitise_column_name(segment_name))
-          )
-        }, map_keys = TRUE)
-    }) %>%
-    purrr::reduce(utils::modifyList) %>%
-    # Then, convert each inner dictionary to a tibble row
-    purrr::map(tibble::as_tibble_row) %>%
-    # And nest each df so that we only have 1 cell of output per sample
-    purrr::map(~tidyr::nest(., {{colname}} := tidyr::everything()))
-}
-
-DEFAULT_PLOT_PARSERS = list(
+DEFAULT_PLOT_PARSERS <- list(
   xy_line = parse_xyline_plot,
   bar_graph = parse_bar_graph
 )
@@ -75,7 +14,7 @@ DEFAULT_PLOT_PARSERS = list(
 #' @noRd
 parse_plots <- function(parsed, plots, plot_parsers) {
   # Merge the default parsers with the user provided ones
-  parsers = purrr::list_modify(DEFAULT_PLOT_PARSERS, !!!plot_parsers)
+  parsers <- purrr::list_modify(DEFAULT_PLOT_PARSERS, !!!plot_parsers)
 
   # Plot data is more complex
   parsed$report_plot_data %>%
@@ -83,11 +22,10 @@ parse_plots <- function(parsed, plots, plot_parsers) {
       # Skip any plot not explicitly in this list, it's impossible to infer
       # what type of plot each is
       if (plot_name %in% plots || is.null(plots)) {
-        parser = parsers[[plot_data$plot_type]]
-        if (!is.null(parser)){
+        parser <- parsers[[plot_data$plot_type]]
+        if (!is.null(parser)) {
           parser(plot_data = plot_data, name = plot_name)
-        }
-        else {
+        } else {
           warning(paste("No known (or provided) parser for a plot of type \"", plot_data$plot_type, "\""))
         }
       }
@@ -97,7 +35,7 @@ parse_plots <- function(parsed, plots, plot_parsers) {
 }
 
 #' List the plot identifiers of all the plots in a given multiqc report
-#' 
+#'
 #' @details The main use for this function is finding the plot identifiers
 #' that you will then pass into the `plots` argument of the [TidyMultiqc::load_multiqc()]
 #' function.
@@ -116,10 +54,10 @@ parse_plots <- function(parsed, plots, plot_parsers) {
 #' filepath <- system.file("extdata", "HG00096/multiqc_data.json", package = "TidyMultiqc")
 #' # This is the actual invocation
 #' list_plots(filepath)
-list_plots <- function(path){
+list_plots <- function(path) {
   jsonlite::read_json(path) %>%
     `$`("report_plot_data") %>%
-    purrr::imap_dfr(function(plot, id){
+    purrr::imap_dfr(function(plot, id) {
       list(
         id = id,
         title = plot$config$title