+++ tests/testthat/test-metadata.R
@@ -20,6 +20,9 @@ test_that("metadata is parsed correctly", {
   expect_true(colnames(report[, 1]) %>% stringr::str_starts("metadata"))
   expect_true(report %>% dplyr::select(last_col()) %>% colnames() %>% stringr::str_starts("general"))
 
+  # We now always put the sample ID first
+  expect_equal(colnames(report[, 1]), "metadata.sample_id")
+
   expect_equal(report$metadata.batch %>% unique(), "P4107")
   expect_equal(report$metadata.sample %>% sort(), c("1001", "1002", "1003", "1004", "1005", "1006"))
 })